Adds good binary type descriptors (#56)
Adds good files for each TL pair and a README that describes the files. As much as possible, I tried to make the representation to be made up of invalid TL pairs. That way even if the implementation is not trying to read the values, it should still fail if it is reading the length incorrectly. 

I split the T6 (Timestamp) and T7 (Symbol) files into two, for implementations that do not want to implement supporting really large VarUInts.